fixed DetectorFactory.java
